Add files via upload
This is an clone project made by me and my 3 teammates with basic html css and javasctipt. We used Local storage as the database and finished this simple clone project. Here i have designed the home page navbar. The reason why we did this assignment is to know how good we are with the basic HTML CSS & JS.